docs: polish up the words in "record.rst"
Fixed the wording of the documentation. It should explains it better and remove the label context of the document, since it is not truly relavent.

Change the wording of the documentation to make it more readable. The `record` type in the documentation is more general than those used only for labels. `record` is an equivalent of `struct` in C++ and it is used to define more complex data structures.